REMOTE warning fixes
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@5486 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/remote/server.cpp
@@ -440,7 +440,7 @@ void SRVR_multi_thread( PORT main_port, USHORT flags)
 			extra_threads = threads_waiting - pending_requests;
 			if (extra_threads < 0) {
 				gds__thread_start(	reinterpret_cast<FPTR_INT_VOID_PTR>(thread),
-									(void*)flags,
+									(void*)(ULONG) flags,
 									THREAD_medium,
 									THREAD_ast,
 									0);
